+++ e2e/client/client.go
@@ -180,11 +180,11 @@ func (c *TestClient) RegisterDataManager(o *DataManagerOptions) {
 		NewPermissions: &asset.NewPermissions{Public: true},
 		Description: &asset.Addressable{
 			Checksum:       "1d55e9c55fa7ad6b6a49ad79da897d58be7ce8b76f92ced4c20f361ba3a0af6e",
-			StorageAddress: "http://somewhere.local/desc",
+			StorageAddress: "http://somewhere.local/desc" + uuid.NewString(),
 		},
 		Opener: &asset.Addressable{
 			Checksum:       "1d55e9c55fa7ad6b6a49ad79da897d58be7ce8b76f92ced4c20f361ba3a0af6e",
-			StorageAddress: "http://somewhere.local/opener",
+			StorageAddress: "http://somewhere.local/opener" + uuid.NewString(),
 		},
 		Type:           "test",
 		LogsPermission: o.LogsPermission,