+++ lib/asset/sql.go
@@ -162,29 +162,12 @@ func (d *DataSample) Value() (driver.Value, error) {
 	return protojson.Marshal(d)
 }
 
-// Value implements the driver.Valuer interface.
-// Returns the JSON-encoded representation of the DataManager.
-func (d *DataManager) Value() (driver.Value, error) {
-	return protojson.Marshal(d)
-}
-
 // Scan implements the sql.Scanner interface.
 // Decodes JSON into a DataSample.
 func (d *DataSample) Scan(value interface{}) error {
 	b, ok := value.([]byte)
 	if !ok {
 		return errors.NewError(errors.ErrByteArray, "cannot scan datasample")
-	}
-
-	return protojson.Unmarshal(b, d)
-}
-
-// Scan implements the sql.Scanner interface.
-// Decodes JSON into the DataManager
-func (d *DataManager) Scan(value interface{}) error {
-	b, ok := value.([]byte)
-	if !ok {
-		return errors.NewError(errors.ErrByteArray, "cannot scan datamanager")
 	}
 
 	return protojson.Unmarshal(b, d)