+++ server/standalone/dbal/datamanager_test.go
@@ -3,12 +3,19 @@ package dbal
 import (
 	"context"
 	"testing"
+	"time"
 
 	"github.com/owkin/orchestrator/lib/common"
 	"github.com/pashagolub/pgxmock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+func makeDataManagerRows() *pgxmock.Rows {
+	return pgxmock.NewRows([]string{"key", "name", "owner", "permissions", "description_address", "description_checksum", "opener_address", "opener_checksum", "type", "creation_date", "logs_permission", "metadata"}).
+		AddRow("key1", "name", "owner", []byte("{}"), "https://example.com/desc", "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855", "https://example.com/chksm", "993b6d90e0ed15d80e7e39c6fb298855d9544420be07faec52935649780e8f19", "", time.Unix(12, 0), []byte("{}"), map[string]string{}).
+		AddRow("key2", "name", "owner", []byte("{}"), "https://example.com/desc", "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855", "https://example.com/chksm", "993b6d90e0ed15d80e7e39c6fb298855d9544420be07faec52935649780e8f19", "", time.Unix(12, 0), []byte("{}"), map[string]string{})
+}
 
 func TestQueryDataManagers(t *testing.T) {
 	mock, err := pgxmock.NewConn()
@@ -18,12 +25,9 @@ func TestQueryDataManagers(t *testing.T) {
 	defer mock.Close(context.Background())
 
 	mock.ExpectBegin()
-
-	rows := pgxmock.NewRows([]string{"asset"}).
-		AddRow([]byte("{}")).
-		AddRow([]byte("{}"))
-
-	mock.ExpectQuery(`select "asset" from "datamanagers"`).WithArgs(uint32(13), 0, testChannel).WillReturnRows(rows)
+	mock.ExpectQuery(`SELECT .* FROM expanded_datamanagers`).
+		WithArgs(testChannel).
+		WillReturnRows(makeDataManagerRows())
 
 	tx, err := mock.Begin(context.Background())
 	require.NoError(t, err)
@@ -48,12 +52,9 @@ func TestPaginatedQueryDataManagers(t *testing.T) {
 	defer mock.Close(context.Background())
 
 	mock.ExpectBegin()
-
-	rows := pgxmock.NewRows([]string{"asset"}).
-		AddRow([]byte("{}")).
-		AddRow([]byte("{}"))
-
-	mock.ExpectQuery(`select "asset" from "datamanagers"`).WithArgs(uint32(2), 0, testChannel).WillReturnRows(rows)
+	mock.ExpectQuery(`SELECT .* FROM expanded_datamanagers`).
+		WithArgs(testChannel).
+		WillReturnRows(makeDataManagerRows())
 
 	tx, err := mock.Begin(context.Background())
 	require.NoError(t, err)